requirements.txt
playwright==1.40.0
anthropic==0.28.0
openai==1.35.0
gigachat==0.1.15
python-dotenv==1.0.0